hmmband: widen cp9b hd_needed/hd_alloced + cur_size to int64_t (brief 147)
Issue-3 fix from brief 144 design: at genome-scale truncated band
derivation, Sum_v(jmax[v]-jmin[v]+1) over ~10^6 CM states with wide
truncated j-bands exceeds 2^31. The int32 cp9b->hd_needed / hd_alloced
(infernal.h) and cur_size (cp9_GrowHDBands) wrap, producing an
undersized hdmin_mem/hdmax_mem allocation and a wild cumulative offset
in hdmin[v] = hdmin_mem + cur_size, so ij2d_bands writes OOB -> SIGSEGV.

Widen the cumulative band-volume integers to int64_t:
  - CP9Bands_t.hd_needed, .hd_alloced (src/infernal.h)
  - cur_size in cp9_GrowHDBands (src/hmmband.c)
  - hd_needed local in cp9_ValidateBands (src/hmmband.c)
Per-state width jbw and per-row index jp stay int (bounded by L+1).
ESL_ALLOC(sizeof(int)*hd_needed) is size_t-promoted -> safe.
esl_vec_ICopy already takes int64_t n -> cp9_CloneBands copy safe.

Non-trunc int path is byte-identical (sum stays under 2^31).

+++ src/hmmband.c
@@ -1358,7 +1358,7 @@ cp9_ValidateBands(CM_t *cm, char *errbuf, CP9Bands_t *cp9b, int i0, int j0, int 
   int sd;           /* minimum d allowed for a state, ex: MP_st = 2, ML_st = 1. etc. */
   int max_sdl_sdr;  /* maximum of StateLeftDelta, StateRightDelta for a state */
   int dn;           /* max_sdl_sdr if do_trunc, else sd */
-  int hd_needed;
+  int64_t hd_needed; /* int64: cumulative band volume can exceed 2^31 at genome-scale truncated (brief 147) */
   int j;
 
 
@@ -1462,8 +1462,8 @@ cp9_GrowHDBands(CP9Bands_t *cp9b, char *errbuf)
 {
   int status;
   int v;
-  int cur_size = 0;
-  int jbw;
+  int64_t cur_size = 0; /* int64: cumulative hd offset can exceed 2^31 at genome-scale truncated (brief 147) */
+  int jbw;              /* per-state j-band width; bounded by L+1, fits int */
 
   /* count size we need for hdmin/hdmax given current jmin, jmax */
   cp9b->hd_needed = 0; /* we'll rewrite this */
